Mudança completa no front e novas implementações de busca

+++ src/components/ListarFornecedores.js
@@ -5,11 +5,12 @@ const ListarFornecedores = () => {
   const [fornecedores, setFornecedores] = useState([]);
   const [editandoFornecedorId, setEditandoFornecedorId] = useState(null);
   const [fornecedorEditado, setFornecedorEditado] = useState({});
+  const [busca, setBusca] = useState('');
 
   useEffect(() => {
     const fetchFornecedores = async () => {
       try {
-        const response = await axios.get('http://localhost:8080/api/fornecedores'); // URL completa
+        const response = await axios.get('http://localhost:8080/api/fornecedores');
         setFornecedores(response.data);
       } catch (error) {
         console.error('Erro ao buscar fornecedores:', error);
@@ -34,7 +35,8 @@ const ListarFornecedores = () => {
       setFornecedores(fornecedores.map(f => (f.id === fornecedorEditado.id ? fornecedorEditado : f)));
       cancelarEdicao();
     } catch (error) {
-      alert('Erro ao salvar fornecedor. Verifique os campos em branco.', error);
+      alert('Erro ao salvar fornecedor. Verifique os campos em branco.');
+      console.error('Erro ao salvar fornecedor:', error);
     }
   };
 
@@ -44,17 +46,44 @@ const ListarFornecedores = () => {
   };
 
   const excluirFornecedor = async (id) => {
-    try {
-      await axios.delete(`http://localhost:8080/api/fornecedores/${id}`);
-      setFornecedores(fornecedores.filter(f => f.id !== id));
-    } catch (error) {
-      console.error('Erro ao excluir fornecedor. Verifique os campos em branco.', error);
+    if (window.confirm('Você tem certeza que deseja excluir este fornecedor?')) {
+      try {
+        await axios.delete(`http://localhost:8080/api/fornecedores/${id}`);
+        setFornecedores(fornecedores.filter(f => f.id !== id));
+      } catch (error) {
+        console.error('Erro ao excluir fornecedor:', error);
+      }
     }
   };
 
+  const handleBuscaChange = (e) => {
+    setBusca(e.target.value);
+  };
+
+  const fornecedoresFiltrados = fornecedores.filter((fornecedor) => {
+    const buscaLower = busca.toLowerCase();
+    return (
+      fornecedor.nome.toLowerCase().includes(buscaLower) ||
+      fornecedor.contato.toLowerCase().includes(buscaLower) ||
+      fornecedor.email.toLowerCase().includes(buscaLower) ||
+      fornecedor.telefone.toLowerCase().includes(buscaLower) ||
+      fornecedor.localizacao.toLowerCase().includes(buscaLower)
+    );
+  });
+
   return (
-    <div>
-      <h1>Listar Fornecedores</h1>
+    <div className='main-container'>
+      <div className='head-container'>
+        <h1>Listar fornecedores: </h1>
+        <div className='search-container'>
+          <input
+            type='text'
+            value={busca}
+            onChange={handleBuscaChange}
+            placeholder='Digite para buscar...'
+          />
+        </div>
+      </div>
       <table>
         <thead>
           <tr>
@@ -67,7 +96,7 @@ const ListarFornecedores = () => {
           </tr>
         </thead>
         <tbody>
-          {fornecedores.map(fornecedor => (
+          {fornecedoresFiltrados.map(fornecedor => (
             <tr key={fornecedor.id}>
               <td>
                 {editandoFornecedorId === fornecedor.id ? (
